Admin dashboard and merged SecurityConfig

+++ src/main/java/com/smart_housing/smart_housing/repository/LandlordRepository.java
@@ -5,6 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LandlordRepository extends JpaRepository<Landlord, Integer> {
+public interface LandlordRepository extends JpaRepository<Landlord, Long> {
     Landlord findByEmail(String email);
+
 }
